Change -listView:heightOfRow: to return a CGFloat instead of an NSInteger. Adjust your code or your row heights will be very screwy!

+++ Classes/AppDelegate.m
@@ -61,7 +61,7 @@
 	return cell;
 }
 
-- (NSInteger)listView:(PXListView*)aListView heightOfRow:(NSInteger)row
+- (CGFloat)listView:(PXListView*)aListView heightOfRow:(NSInteger)row
 {
 	return 50;
 }
