+++ Classes/PXListViewDelegate.h
@@ -15,6 +15,6 @@
 @required
 - (NSInteger)numberOfRowsInListView:(PXListView*)aListView;
 - (PXListViewCell*)listView:(PXListView*)aListView cellForRow:(NSInteger)row;
-- (NSInteger)listView:(PXListView*)aListView heightOfRow:(NSInteger)row;
+- (CGFloat)listView:(PXListView*)aListView heightOfRow:(NSInteger)row;
 
 @end
